@@ -2,67 +2,116 @@
 
 namespace Edg\Erp\Cron\API;
 
+use Edg\Erp\Helper\Data;
+use Exception;
 use Magento\Catalog\Api\Data\ProductInterface;
+use Magento\Catalog\Api\ProductRepositoryInterface;
+use Magento\CatalogInventory\Api\StockStatusRepositoryInterface;
+use Magento\CatalogInventory\Api\StockItemRepositoryInterface;
 use Magento\Catalog\Model\Product\Attribute\Source\Status;
+use Magento\Catalog\Model\Product\TierPriceManagement;
+use Magento\CatalogRule\Model\Rule\Job;
+use Magento\Framework\App\Cache;
 use Magento\Framework\App\Config\ConfigResource\ConfigInterface;
 use Magento\Framework\App\Filesystem\DirectoryList;
-use Magento\Framework\Mail\Message;
+use Magento\Framework\Exception\CouldNotSaveException;
+use Magento\Framework\Exception\NoSuchEntityException;
+use Magento\Framework\Mail\Template\TransportBuilder;
+use Magento\Framework\Exception\FileSystemException;
+use Magento\Framework\Logger\Monolog;
+use Magento\Store\Model\StoreManager;
 
 class ArticleInfo extends AbstractCron
 {
     /**
-     * @var \Magento\Catalog\Api\ProductRepositoryInterface
+     * @var ProductRepositoryInterface
      */
-    protected $productRepository;
+    protected ProductRepositoryInterface $productRepository;
 
     /**
-     * @var \Magento\CatalogInventory\Api\StockRegistryInterface
+     * @var StockStatusRepositoryInterface
      */
-    protected $stockRegistry;
+    protected StockStatusRepositoryInterface $stockStatusRepository;
 
     /**
-     * @var \Magento\Catalog\Model\Product\TierPriceManagement
+     * @var StockItemRepositoryInterface
      */
-    protected $tierpriceManagement;
+    protected StockItemRepositoryInterface $stockItemRepository;
 
     /**
-     * @var \Magento\CatalogRule\Model\Rule\Job
+     * @var TierPriceManagement
      */
-    protected $catalogRuleJob;
+    protected TierPriceManagement $tierpriceManagement;
 
     /**
-     * @var \Magento\Framework\App\Cache
+     * @var TransportBuilder
      */
-    protected $cache;
-    protected $fieldsToSkip = [
+    protected TransportBuilder $transportBuilder;
+
+    /**
+     * @var Job
+     */
+    protected Job $catalogRuleJob;
+
+    /**
+     * @var Cache
+     */
+    protected Cache $cache;
+
+    /**
+     * @var string[]
+     */
+    protected array $fieldsToSkip = [
         'sku'
     ];
     /**
      * Array of skus to call on
      */
-    protected $_skuList = [];
+    protected array $_skuList = [];
 
-    protected $apiMessages = [];
+    /**
+     * @var array
+     */
+    protected array $apiMessages = [];
 
+    /**
+     * @param Data $helper
+     * @param DirectoryList $directoryList
+     * @param Monolog $monolog
+     * @param ConfigInterface $config
+     * @param TransportBuilder $transportBuilder
+     * @param StoreManager $storeManager
+     * @param ProductRepositoryInterface $productRepository
+     * @param StockStatusRepositoryInterface $stockStatusRepository
+     * @param StockItemRepositoryInterface $stockItemRepository
+     * @param TierPriceManagement $tierPriceManagement
+     * @param Job $catalogRuleJob
+     * @param Cache $cache
+     * @param array $settings
+     * @throws FileSystemException
+     */
     public function __construct(
-        \Edg\Erp\Helper\Data $helper,
+        Data $helper,
         DirectoryList $directoryList,
+        Monolog $monolog,
         ConfigInterface $config,
-        Message $message,
-        \Magento\Store\Model\StoreManager $storeManager,
-        \Magento\Catalog\Api\ProductRepositoryInterface $productRepository,
-        \Magento\CatalogInventory\Api\StockRegistryInterface $stockRegistryInterface,
-        \Magento\Catalog\Model\Product\TierPriceManagement $tierPriceManagement,
-        \Magento\CatalogRule\Model\Rule\Job $catalogRuleJob,
-        \Magento\Framework\App\Cache $cache,
-        $settings = []
+        TransportBuilder $transportBuilder,
+        StoreManager $storeManager,
+        ProductRepositoryInterface $productRepository,
+        StockStatusRepositoryInterface $stockStatusRepository,
+        StockItemRepositoryInterface $stockItemRepository,
+        TierPriceManagement $tierPriceManagement,
+        Job $catalogRuleJob,
+        Cache $cache,
+        array $settings = []
     ) {
         $this->productRepository = $productRepository;
-        $this->stockRegistry = $stockRegistryInterface;
+        $this->stockStatusRepository = $stockStatusRepository;
+        $this->stockItemRepository = $stockItemRepository;
         $this->tierpriceManagement = $tierPriceManagement;
         $this->catalogRuleJob = $catalogRuleJob;
         $this->cache = $cache;
-        parent::__construct($helper, $directoryList, $config, $message, $storeManager, $settings);
+        parent::__construct($helper, $directoryList, $monolog, $config, $transportBuilder, $storeManager, $settings);
     }
 
     public function execute()
@@ -76,7 +125,7 @@ class ArticleInfo extends AbstractCron
                 $this->catalogRuleJob->applyAll();
                 $this->cache->remove('catalog_rules_dirty');
                 $this->moduleLog(__METHOD__ . '(); - ' . __('The rules have been applied.'), true);
-            } catch (\Exception $exception) {
+            } catch (Exception $exception) {
                 $this->moduleLog(__METHOD__ . '(); - ' . __('Unable to apply rules.'), true);
                 $this->moduleLog(__METHOD__ . '(); - ' . $exception->getMessage(), true);
             }
@@ -88,7 +137,7 @@ class ArticleInfo extends AbstractCron
         $this->apiMessages = $response;
     }
 
-    protected function processProductUpdates()
+    protected function processProductUpdates(): array
     {
         $messages = [];
 
@@ -103,7 +152,7 @@ class ArticleInfo extends AbstractCron
         if (sizeof($this->_skuList) == 0) {
             $this->_skuList = $this->getSkuArray();
         }
-        
+
         if ($this->_skuList) {
             $results = $client->pullArticleInfo($this->_skuList);
 
@@ -200,12 +249,12 @@ class ArticleInfo extends AbstractCron
 
             try {
                 $this->productRepository->save($product);
-            } catch (\Exception $e) {
+            } catch (Exception $e) {
                 $this->moduleLog(__METHOD__ . '() - WARNING: Errors attemping to update product with sku "' . $productdata['sku'] . '": ' . $e->getMessage());
             }
-        } catch (\Magento\Framework\Exception\NoSuchEntityException $e) {
+        } catch (NoSuchEntityException $e) {
             $this->moduleLog(__METHOD__ . '() - WARNING: Product "' . $sku . '" failed to load.');
-        } catch (\Exception $e) {
+        } catch (Exception $e) {
             $this->moduleLog(__METHOD__ . '() - WARNING: Errors attemping to update product with sku "' . $productdata['sku'] . '": ' . $e->getMessage());
             $this->moduleLog($e->getTraceAsString(), true);
         }
@@ -218,7 +267,7 @@ class ArticleInfo extends AbstractCron
      * @param \Edg\ErpService\DataModel\ArticleInfo $article
      * @return $this
      */
-    protected function updateProductStatus(ProductInterface $product, \Edg\ErpService\DataModel\ArticleInfo $article)
+    protected function updateProductStatus(ProductInterface $product, \Edg\ErpService\DataModel\ArticleInfo $article): ArticleInfo
     {
         $orderable = $article->getOrderable();
 
@@ -257,34 +306,39 @@ class ArticleInfo extends AbstractCron
      * @param ProductInterface $product
      * @param \Edg\ErpService\DataModel\ArticleInfo $article
      * @return $this
+     * @throws Exception
      */
     protected function updateProductStockSettings(
         ProductInterface $product,
         \Edg\ErpService\DataModel\ArticleInfo $article
-    ) {
-        $stock = $this->stockRegistry->getStockItem($product->getId());
-        if ($stock->getItemId()) {
+    ): ArticleInfo
+    {
+        $stockItem = $this->stockItemRepository->get($product->getExtensionAttributes()->getStockItem()->getItemId());
+        $stockStatus = $this->stockStatusRepository->get($stockItem->getItemId());
+
+        if ($stockItem && $stockStatus) {
             $sku = $product->getSku();
             $newQty = $article->getAvailable();
             $newBackorders = ($article->getBackorder() == \Edg\ErpService\DataModel\ArticleInfo::BACKORDER_TRUE) ? 1 : 0;
-            $oldBackorders = $stock->getBackorders();
+            $oldBackorders = $stockItem->getBackorders();
 
             $inStock = ($newQty > 0 || $newBackorders == 1) ? 1 : 0;
 
-            $productQtyAndStock = $product->getQuantityAndStockStatus();
+            $productQty = $stockStatus->getQty();
+            $productStockStatus = $stockStatus->getStockStatus();
 
-            if ($stock->getUseConfigBackorders() != 0) {
-                $stock->setUseConfigBackorders(false);
+            if ($stockItem->getUseConfigBackorders() != 0) {
+                $stockItem->setUseConfigBackorders(false);
             }
 
-            if ($stock->getQty() != $newQty) {
-                $stock->setQty($newQty);
-                $productQtyAndStock['qty'] = $newQty;
+            if ($stockStatus->getQty() != $newQty) {
+                $stockItem->setQty($newQty);
+                $productQty = $newQty;
             }
 
-            if ($stock->getIsInStock() != $inStock) {
-                $stock->setIsInStock($inStock);
-                $productQtyAndStock['is_in_stock'] = $inStock;
+            if ($stockItem->getIsInStock() != $inStock) {
+                $stockItem->setIsInStock($inStock);
+                $productStockStatus = $inStock;
             }
 
             $this->moduleLog(__METHOD__ . sprintf(
@@ -296,7 +350,7 @@ class ArticleInfo extends AbstractCron
             ), true);
 
             if ($newBackorders != $oldBackorders) {
-                $stock->setBackorders($newBackorders);
+                $stockItem->setBackorders($newBackorders);
                 $this->moduleLog(
                     __METHOD__ . ': Product #' . $sku . ': Backorder status "' . $oldBackorders . '" updated to "' . $newBackorders . '".',
                     true
@@ -304,9 +358,11 @@ class ArticleInfo extends AbstractCron
             } else {
                 $this->moduleLog(__METHOD__ . ': Product #' . $sku . ': No need for backorder update.', true);
             }
+            $stockItem->setQty($productQty);
+            $stockStatus->setStockStatus($productStockStatus);
 
-            $this->stockRegistry->updateStockItemBySku($sku, $stock);
-            $product->setQuantityAndStockStatus($productQtyAndStock);
+            $stockItem->save();
+            $this->stockStatusRepository->save($stockStatus);
         } else {
             $this->moduleLog(__METHOD__ . '() - WARNING: could not load stock model for product "' . $product->getSku() . '"');
         }
@@ -317,8 +373,8 @@ class ArticleInfo extends AbstractCron
      * @param ProductInterface $product
      * @param \Edg\ErpService\DataModel\ArticleInfo $article
      * @return $this|bool
-     * @throws \Magento\Framework\Exception\CouldNotSaveException
-     * @throws \Magento\Framework\Exception\InputException
+     * @throws CouldNotSaveException
+     * @throws NoSuchEntityException
      */
     protected function updateProductPriceTiers(
         ProductInterface $product,
@@ -382,7 +438,7 @@ class ArticleInfo extends AbstractCron
                     $price,
                     $qty
                 );
-            } catch (\Exception $e) {
+            } catch (Exception $e) {
                 $this->moduleLog(__METHOD__ . '() - ERROR: could not save price tier for product #' . $product->getSku());
                 $this->moduleLog($e->getMessage());
             }
@@ -395,9 +451,10 @@ class ArticleInfo extends AbstractCron
         return $this;
     }
 
-    protected function updateProductTaxClass(ProductInterface $product, \Edg\ErpService\DataModel\ArticleInfo $article)
+    protected function updateProductTaxClass(ProductInterface $product, \Edg\ErpService\DataModel\ArticleInfo $article): ArticleInfo
     {
         $map = $this->helper->getTaxClassMapping();
+
         $rate = (string)(float)str_replace(',', '.', $article->getBtw());
 
         if (array_key_exists($rate, $map)) {
@@ -407,7 +464,9 @@ class ArticleInfo extends AbstractCron
                 $map[$rate],
                 $rate
             ), true);
+
             $product->setTaxClassId($map[$rate]);
+
         } else {
             $this->moduleLog(__METHOD__ . ': Product #' . $product->getSku() . ': WARNING: No matching tax_class_id for rate "' . $rate . '",');
         }
@@ -419,7 +478,7 @@ class ArticleInfo extends AbstractCron
      * @param String $sku
      * @return $this
      */
-    public function addSkuToSync($sku)
+    public function addSkuToSync(string $sku): ArticleInfo
     {
         $this->_skuList[] = $sku;
         return $this;
@@ -428,7 +487,7 @@ class ArticleInfo extends AbstractCron
     /**
      * @return array
      */
-    public function getApiMessages()
+    public function getApiMessages(): array
     {
         return $this->apiMessages;
     }
