@@ -2,39 +2,98 @@
 
 namespace Edg\Erp\Cron\API;
 
+use Edg\Erp\Helper\Data;
+use Edg\ErpService\DataModel\OrderStatus;
+use Exception;
 use Magento\Framework\App\Config\ConfigResource\ConfigInterface;
 use Magento\Framework\App\Filesystem\DirectoryList;
-use Magento\Framework\Mail\Message;
+use Magento\Framework\DB\Transaction;
+use Magento\Framework\Exception\FileSystemException;
+use Magento\Framework\Exception\LocalizedException;
+use Magento\Framework\Exception\NoSuchEntityException;
+use Magento\Framework\Logger\Monolog;
+use Magento\Framework\Mail\Template\TransportBuilder;
+use Magento\Sales\Model\Order;
+use Magento\Sales\Model\Order\Email\Sender\ShipmentSender;
+use Magento\Sales\Model\Order\Shipment;
+use Magento\Sales\Model\Order\ShipmentFactory;
+use Magento\Sales\Model\OrderFactory;
+use Magento\Store\Model\StoreManager;
+use Monolog\Logger;
 
 class OrderStatusImport extends AbstractCron
 {
     const TRACKING_TITLE = 'Track & Trace';
     const CARRIER_CODE = 'PostNL';
 
-    protected $orderFactory;
-    protected $shipmentFactory;
-    protected $transaction;
-    protected $shipmentMailer;
+    /**
+     * @var OrderFactory
+     */
+    protected OrderFactory $orderFactory;
 
+    /**
+     * @var ShipmentFactory
+     */
+    protected ShipmentFactory $shipmentFactory;
+
+    /**
+     * @var Transaction
+     */
+    protected Transaction $transaction;
+
+    /**
+     * @var ShipmentSender
+     */
+    protected ShipmentSender $shipmentMailer;
+
+    /**
+     * @var Monolog
+     */
+    protected Monolog $monolog;
+
+    /**
+     * @var TransportBuilder
+     */
+    protected TransportBuilder $transportBuilder;
+
+    /**
+     * @param Data $helper
+     * @param DirectoryList $directoryList
+     * @param Monolog $monolog
+     * @param ConfigInterface $config
+     * @param TransportBuilder $transportBuilder
+     * @param StoreManager $storeManager
+     * @param OrderFactory $orderFactory
+     * @param ShipmentFactory $shipmentFactory
+     * @param Transaction $transaction
+     * @param ShipmentSender $sender
+     * @param array $settings
+     * @throws FileSystemException
+     */
     public function __construct(
-        \Edg\Erp\Helper\Data $helper,
+        Data $helper,
         DirectoryList $directoryList,
+        Monolog $monolog,
         ConfigInterface $config,
-        Message $message,
-        \Magento\Store\Model\StoreManager $storeManager,
-        \Magento\Sales\Model\OrderFactory $orderFactory,
-        \Magento\Sales\Model\Order\ShipmentFactory $shipmentFactory,
-        \Magento\Framework\DB\Transaction $transaction,
-        \Magento\Sales\Model\Order\Email\Sender\ShipmentSender $sender,
+        TransportBuilder $transportBuilder,
+        StoreManager $storeManager,
+        OrderFactory $orderFactory,
+        ShipmentFactory $shipmentFactory,
+        Transaction $transaction,
+        ShipmentSender $sender,
         array $settings = []
     ) {
-        parent::__construct($helper, $directoryList, $config, $message, $storeManager, $settings);
+        parent::__construct($helper, $directoryList, $monolog, $config, $transportBuilder, $storeManager, $settings);
         $this->orderFactory = $orderFactory;
         $this->shipmentFactory = $shipmentFactory;
         $this->transaction = $transaction;
         $this->shipmentMailer = $sender;
     }
 
+    /**
+     * @throws NoSuchEntityException
+     * @throws LocalizedException
+     */
     public function execute()
     {
         if ($this->helper->isPullOrdersEnabled()) {
@@ -42,11 +101,16 @@ class OrderStatusImport extends AbstractCron
         }
     }
 
+    /**
+     * @throws NoSuchEntityException
+     * @throws LocalizedException
+     * @throws Exception
+     */
     protected function prepareImport()
     {
         $date = date("Y_m_d");
-        $this->addLogStreamToServiceLogger($this->_importDir . DIRECTORY_SEPARATOR . "log_{$date}.log");
-        $this->serviceLog('Start import');
+        $stream = $this->addLogStreamToServiceLogger($this->_importDir . DIRECTORY_SEPARATOR . "log_{$date}.log");
+        $this->serviceLog($stream, 'Start import');
 
         $shippedStatus = $this->helper->getOrderImportStatusAfterShipping();
         $environment = $this->helper->getEnvironmentTag();
@@ -61,19 +125,19 @@ class OrderStatusImport extends AbstractCron
             }
             $orders = $response->getOrders();
             foreach ($orders as $order) {
-                $this->serviceLog('starting import for order with order number: #' . $order->getOrderNumber());
+                $this->serviceLog($stream,'starting import for order with order number: #' . $order->getOrderNumber());
                 $orderIncrementId = $this->formatIncrementId($order->getOrderNumber());
                 try {
                     $magentoOrder = $this->orderFactory->create()->loadByIncrementId($orderIncrementId);
 
                     if (!$magentoOrder->getId()) {
-                        $this->serviceLog("Order " . $orderIncrementId . " can not be found\n");
+                        $this->serviceLog($stream, "Order " . $orderIncrementId . " can not be found\n");
                         $this->moduleLog('WARNING: Order #' . $orderIncrementId . ' cannot be found.');
                         continue;
                     }
 
-                    if ($order->getOrderStatus() == \Edg\ErpService\DataModel\OrderStatus::STATUS_NOT_SHIPPED) {
-                        $this->serviceLog("Order " . $orderIncrementId . " is not shipped, no need to update \n");
+                    if ($order->getOrderStatus() == OrderStatus::STATUS_NOT_SHIPPED) {
+                        $this->serviceLog($stream, "Order " . $orderIncrementId . " is not shipped, no need to update \n");
                         continue;
                     }
 
@@ -88,10 +152,10 @@ class OrderStatusImport extends AbstractCron
                         ];
                     }
 
-                    $this->serviceLog("Creating shipment for order #" . $magentoOrder->getIncrementId() . " with items " . print_r($itemsToShip,
+                    $this->serviceLog($stream, "Creating shipment for order #" . $magentoOrder->getIncrementId() . " with items " . print_r($itemsToShip,
                             true));
 
-                    /** @var \Magento\Sales\Model\Order\Shipment $shipment */
+                    /** @var Shipment $shipment */
                     $shipment = $this->shipmentFactory->create($magentoOrder, $itemsToShip, $tracks);
                     $shipment->register();
 
@@ -103,6 +167,7 @@ class OrderStatusImport extends AbstractCron
                     $this->saveShipment($shipment);
 
                     $this->moduleLog(sprintf('Created shipment #%s (id %s) for order #%s (id %s)',
+                        $shipment->getIncrementId(),
                         $shipment->getIncrementId(),
                         $shipment->getId(),
                         $magentoOrder->getIncrementId(),
@@ -116,17 +181,18 @@ class OrderStatusImport extends AbstractCron
                         $this->moduleLog('Bypassing shipment email notification for #' . $magentoOrder->getIncrementId());
                     }
 
-                    $this->serviceLog("Order #" . $magentoOrder->getIncrementId() . " saved successfully with status " . $magentoOrder->getStatus());
-                } catch (\Exception $e) {
+                    $this->serviceLog($stream, "Order #" . $magentoOrder->getIncrementId() . " saved successfully with status " . $magentoOrder->getStatus());
+                } catch (Exception $e) {
 
                     if (stripos($e->getMessage(),
                             'Expectation failed') !== false
-                    ) { //PHPUnit exception; must not be catched
+                    ) {
+                        //PHPUnit exception; must not be catched
                         throw $e;
                     }
 
-                    $this->serviceLog("Order " . $orderIncrementId . " failed import", \Zend\Log\Logger::ERR);
-                    $this->serviceLog($e->getMessage(), \Zend\Log\Logger::ERR);
+                    $this->serviceLog($stream, "Order " . $orderIncrementId . " failed import", Logger::ERROR);
+                    $this->serviceLog($stream, $e->getMessage(), Logger::ERROR);
 
                     $this->moduleLog('WARNING: Order ' . $orderIncrementId . ' failed import');
                     $this->moduleLog($e->getMessage());
@@ -145,8 +211,9 @@ class OrderStatusImport extends AbstractCron
      *
      * @param $id
      * @return string
+     * @throws Exception
      */
-    protected function formatIncrementId($id)
+    protected function formatIncrementId($id): string
     {
         $pattern = \Magento\SalesSequence\Model\Sequence::DEFAULT_PATTERN;
         $pos1 = strpos($pattern, '%');
@@ -156,20 +223,21 @@ class OrderStatusImport extends AbstractCron
 
         $newId = sprintf($newPattern, $id);
         if (strlen($newId) != strlen($id)) {
-            $this->serviceLog(sprintf('Transforming Order increment id from %s to %s', $id, $newId));
+            $this->serviceLog(null, sprintf('Transforming Order increment id from %s to %s', $id, $newId));
         }
         return $newId;
     }
 
     /**
-     * @param \Magento\Sales\Model\Order $magentoOrder
-     * @param \Edg\ErpService\DataModel\OrderStatus $importData
+     * @param Order $magentoOrder
+     * @param OrderStatus $importData
      * @return array
      */
     protected function getItemsToShip(
-        \Magento\Sales\Model\Order $magentoOrder,
-        \Edg\ErpService\DataModel\OrderStatus $importData
-    ) {
+        Order $magentoOrder,
+        OrderStatus $importData
+    ): array
+    {
         $items = [];
 
         foreach ($importData->getOrderRows() as $orderrow) {
@@ -221,10 +289,11 @@ class OrderStatusImport extends AbstractCron
     /**
      * Save shipment and order in one transaction
      *
-     * @param \Magento\Sales\Model\Order\Shipment $shipment
+     * @param Shipment $shipment
      * @return $this
+     * @throws Exception
      */
-    protected function saveShipment($shipment)
+    protected function saveShipment(Shipment $shipment): OrderStatusImport
     {
         $shipment->getOrder()->setIsInProcess(true);
         $transaction = $this->transaction;
