@@ -2,29 +2,59 @@
 
 namespace Edg\Erp\Cron\API;
 
-use Magento\Catalog\Api\Data\ProductInterface;
-use Magento\Catalog\Model\Product\Attribute\Source\Status;
+use Edg\Erp\Helper\Data;
+use Magento\Catalog\Api\ProductRepositoryInterface;
+use Magento\Catalog\Model\Product\TierPriceManagement;
+use Magento\CatalogInventory\Api\StockStatusRepositoryInterface;
+use Magento\CatalogInventory\Api\StockItemRepositoryInterface;
+use Magento\CatalogRule\Model\Rule\Job;
+use Magento\Framework\App\Cache;
 use Magento\Framework\App\Config\ConfigResource\ConfigInterface;
 use Magento\Framework\App\Filesystem\DirectoryList;
-use Magento\Framework\Mail\Message;
+use Magento\Framework\Exception\FileSystemException;
+use Magento\Framework\Mail\Template\TransportBuilder;
+use Magento\Framework\Logger\Monolog;
+use Magento\Store\Model\StoreManager;
 
 class ArticleInfoUpdate extends ArticleInfo
 {
+    /**
+     * @var TransportBuilder
+     */
+    protected TransportBuilder $transportBuilder;
 
+    /**
+     * @param Data $helper
+     * @param DirectoryList $directoryList
+     * @param Monolog $monolog
+     * @param ConfigInterface $config
+     * @param TransportBuilder $transportBuilder
+     * @param StoreManager $storeManager
+     * @param ProductRepositoryInterface $productRepository
+     * @param StockStatusRepositoryInterface $stockStatusRepository
+     * @param StockItemRepositoryInterface $stockItemRepository
+     * @param TierPriceManagement $tierPriceManagement
+     * @param Job $catalogRuleJob
+     * @param Cache $cache
+     * @param array $settings
+     * @throws FileSystemException
+     */
     public function __construct(
-        \Edg\Erp\Helper\Data $helper,
+        Data $helper,
         DirectoryList $directoryList,
+        Monolog $monolog,
         ConfigInterface $config,
-        Message $message,
-        \Magento\Store\Model\StoreManager $storeManager,
-        \Magento\Catalog\Api\ProductRepositoryInterface $productRepository,
-        \Magento\CatalogInventory\Api\StockRegistryInterface $stockRegistryInterface,
-        \Magento\Catalog\Model\Product\TierPriceManagement $tierPriceManagement,
-        \Magento\CatalogRule\Model\Rule\Job $catalogRuleJob,
-        \Magento\Framework\App\Cache $cache,
-        $settings = []
+        TransportBuilder $transportBuilder,
+        StoreManager $storeManager,
+        ProductRepositoryInterface $productRepository,
+        StockStatusRepositoryInterface $stockStatusRepository,
+        StockItemRepositoryInterface $stockItemRepository,
+        TierPriceManagement $tierPriceManagement,
+        Job $catalogRuleJob,
+        Cache $cache,
+        array $settings = []
     ) {
-        parent::__construct($helper, $directoryList, $config, $message, $storeManager, $productRepository,$stockRegistryInterface, $tierPriceManagement, $catalogRuleJob, $cache,$settings);
+        parent::__construct($helper, $directoryList, $monolog, $config, $transportBuilder, $storeManager, $productRepository, $stockStatusRepository, $stockItemRepository, $tierPriceManagement, $catalogRuleJob, $cache, $settings);
     }
 
     /**
